props removal
props removal

+++ BlazorWASM/RWebComponents/Controls/Canvas/DOMMatrixInit.cs
@@ -3,16 +3,18 @@
 
 namespace RWebComponents.Controls.Canvas;
 
+
+
 public class DOMMatrixInit : DOMMatrix2DInit {
     public bool? is2D {get; set; }
-    public int? m13 { set; get; }
-    public int? m14{ set; get; }
-    public int? m23{ set; get; }
-    public int? m24{ set; get; }
-    public int? m31{ set; get; }
-    public int? m32{ set; get; }
-    public int? m33{ set; get; }
-    public int? m34{ set; get; }
-    public int? m43{ set; get; }
-    public int? m44{ set; get; }
+    public double? m13 { set; get; }
+    public double? m14{ set; get; }
+    public double? m23{ set; get; }
+    public double? m24{ set; get; }
+    public double? m31{ set; get; }
+    public double? m32{ set; get; }
+    public double? m33{ set; get; }
+    public double? m34{ set; get; }
+    public double? m43{ set; get; }
+    public double? m44{ set; get; }
 }
